SCRUM-6 refactoring imports
