+++ src/app/main.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { createRoot } from "react-dom/client"
 import { Provider } from "react-redux"
 import App from "./App"
-import { store } from "../app/store"
+import { store } from "@/app/store"
 import "../app/index.css"
 
 const container = document.getElementById("root")
