+++ vite.config.ts
@@ -1,9 +1,15 @@
 import { defineConfig } from "vitest/config"
 import react from "@vitejs/plugin-react"
+import path from "node:path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src'), // Теперь @ ссылается на src
+    },
+  },
   server: {
     open: true,
   },
